+++ CHANGELOG.md
@@ -2,10 +2,24 @@
 
 ## Development
 
+## 0.13.0
+
 ### Features
 
+* added Kubernetes API objects:
+  * `Event`
+  * `ResourceQuota`
+  * `ServiceAccount`
+  * `ThirdPartyResource`
+  * `PetSet`
+  * `HorizontalPodAutoscaler`
 * `KubeConfig` learned to handle empty or missing user configuration
 * `HTTPClient` learned to create an HTTP session with no authentication
+* `Deployment` learned to report itself ready
+
+### Bug fixes
+
+* `WatchQuery` learned to query against non-v1 API objects
 
 ## 0.12.0